refactor: Consolidate into single file (remove utils dir)
